auth/utils/auth.py
import jwt
import bcrypt

import uuid
from datetime import timedelta, datetime, UTC

from api.exceptions.auth import (
    TokenTypeException,
)
from core.config import settings, TOKEN_TYPE_FIELD


def encode_jwt(
    payload: dict,
    private_key: str = settings.auth.private_key_path.read_text(),
    algorithm: str = settings.auth.algorithm,
    expire_timedelta: timedelta | None = None,
    expire_minutes: int = settings.auth.access_token_expire_minutes,
):
    to_encode = payload.copy()
    now = datetime.now(UTC)
    if expire_timedelta:
        expire = now + expire_timedelta
    else:
        expire = now + timedelta(minutes=expire_minutes)
    to_encode.update(exp=expire, iat=now)
    encoded = jwt.encode(
        payload=to_encode,
        key=private_key,
        algorithm=algorithm,
    )
    return encoded


def decode_jwt(
    token: str | bytes,
    public_key: str = settings.auth.public_key_path.read_text(),
    algorithm: str = settings.auth.algorithm,
):
    decoded = jwt.decode(token, public_key, algorithms=[algorithm])
    return decoded


def hash_password(password: str) -> bytes:
    salt = bcrypt.gensalt()
    pwd_bytes: bytes = password.encode()
    return bcrypt.hashpw(pwd_bytes, salt)


def validate_password(
    password: str,
    hashed_password: bytes,
) -> bool:
    return bcrypt.checkpw(
        password=password.encode(),
        hashed_password=hashed_password,
    )


def validate_token_type(
    payload: dict,
    got_token_type: str,
) -> bool:
    token_type = payload.get(TOKEN_TYPE_FIELD)
    if got_token_type != token_type:
        raise TokenTypeException
    return True
