pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standart"], version = "^0.35.0"}
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.10.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.42"}
asyncpg = "^0.30.0"
alembic = "^1.16.4"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
orjson = "^3.11.1"
loguru = "^0.7.3"
user-agents = "^2.2.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

